manager: Force AllowSystemInstallation to true for administrators
This is a downstream workaround for
https://phabricator.endlessm.com/T27854 while we work on setting the
appropriate `AllowSystemInstallation` value by default in
administrators’ profiles (but not the profiles of normal users).

It queries accounts-service for each user when loading their app
filter, and overrides the value of `AllowSystemInstallation` for
administrators.

The test cases need to be updated for the new D-Bus call. This includes
a new test case for the changed downstream behaviour.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T27854