po: Update translations from downstream in Endless
We’ll drop the Transifex translation machinery downstream after this, so
this is the last time the translations will be synced.

I have not verified any of these translations, but am upstreaming them
on the premise that some strings are better than no strings.

This commit was produced by:
 1. `ninja malcontent-update-po`
 2. Copy in po files from downstream
 3. `ninja malcontent-update-po` again
 4. Manually drop header-only changes
 5. Manually drop changes to `translator-credits`
 6. Manually drop some conflicted strings from `id.po` (prefer the
    upstream versions)
 7. Manually merge headers

There are no more `help/` translations to upstream.

Signed-off-by: Philip Withnall <withnall@endlessm.com>